customer/serializers.py
from rest_framework import serializers

from customer.models import Customer


class CustomerSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Customer
        fields = ['id', 'url', 'first_name', 'last_name', 'email', 'gender']
        
        
    def validate_email(self, value):
        if Customer.objects.filter(email=value).exists():
            raise serializers.ValidationError('Пользователь с таким email уже существует.')
        return value
    
